Move each class and module to its own file
This should make it easier to find each class.
