+++ lib/middleman-syntax/extension.rb
@@ -1,4 +1,7 @@
 require 'rouge'
+require 'middleman-syntax/highlighter'
+require 'middleman-syntax/redcarpet_code_renderer'
+require 'middleman-syntax/haml_monkey_patch'
 
 module Middleman
   module Syntax
@@ -59,54 +62,6 @@ module Middleman
           content = content.encode(Encoding::UTF_8)
 
           concat_content Middleman::Syntax::Highlighter.highlight(content, language).html_safe
-        end
-      end
-    end
-
-    # A mixin for the Redcarpet Markdown renderer that will highlight
-    # code.
-    module RedcarpetCodeRenderer
-      def block_code(code, language)
-        Middleman::Syntax::Highlighter.highlight(code, language)
-      end
-    end
-
-    module Highlighter
-      mattr_accessor :options
-
-      # A helper module for highlighting code
-      def self.highlight(code, language=nil, opts={})
-        lexer = Rouge::Lexer.find_fancy(language, code) || Rouge::Lexers::PlainText
-
-        highlighter_options = options.to_h.merge(opts)
-        highlighter_options[:css_class] = [ highlighter_options[:css_class], lexer.tag ].join(' ')
-        lexer_options = highlighter_options.delete(:lexer_options)
-
-        formatter = Rouge::Formatters::HTML.new(highlighter_options)
-        formatter.format(lexer.lex(code, options.lexer_options))
-      end
-    end
-  end
-end
-
-# If Haml is around, define a :code filter that can be used to more conveniently output highlighted code.
-if defined? Haml
-  module Haml
-    module Filters
-      module Code
-        include Base
-
-        def render(code)
-          code = code.encode(Encoding::UTF_8)
-
-          # Allow language to be specified via a special comment like:
-          #  # lang: ruby
-          if code.lines.first =~ /\A\s*#\s*lang:\s*(\w+)$/
-            language = $1
-            code = code.lines.to_a[1..-1].join # Strip first line
-          end
-
-          Middleman::Syntax::Highlighter.highlight(code, language)
         end
       end
     end
