+++ lib/middleman-syntax/redcarpet_code_renderer.rb
@@ -1,0 +1,11 @@
+module Middleman
+  module Syntax
+    # A mixin for the Redcarpet Markdown renderer that will highlight
+    # code.
+    module RedcarpetCodeRenderer
+      def block_code(code, language)
+        Middleman::Syntax::Highlighter.highlight(code, language)
+      end
+    end
+  end
+end
